+++ EzDeal/API JAVA/src/main/java/com/myorg/ezdeal/repository/CuentaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CuentaRepository extends JpaRepository<Cuenta, Integer> {
+public interface CuentaRepository extends JpaRepository<Cuenta, Long> {
 
     Optional<Cuenta> findByEmail(String email);
     Cuenta findByNombreUsuarioAndContrasena(String nombreUsuario, String contrasena);
